Change to using `isot` formated astropy.time.Time for history (#104)
test_history fails if asdf-astropy is installed. This is because asdf-astropy needed to alter how it wrote times to roundtrip values. Currently only `isot`, `jyear`,  and `byear` formats are written to strings as needed for the tests in question as those are the only ones which can be recovered in the round-trip successfully.

Add asdf-astropy requirement

asdf-astropy is already required by jwst, this adds that here so that we explicitly catch asdf-astropy related issues within stdatamodels. Moreover the astropy fallback for asdf is deprecated and will be removed in October 2023. This pin will prevent further dependence on that astropy fall back by accident.

Add changelog

Co-authored-by: William Jamieson <wjamieson@stsci.edu>