Resolves JP-2798 improves memory usage of datamodel (#109)
* improve memory usage of datamodel

Remove the self-reference (self._ctx = self) and other
difficult to garbage collect patterns to help python clean
up when a process allocates and saves several models.

DataModel contains a self reference that makes it difficult
to collect. This PR converts DataModel._ctx to a property
to work around this.

The use of `validate` to generate the hdulist suffers from
leakage as validators are cached resulting in references to
the hdulist remaining after a model deleted. This PR changes
the hdulist ref to a weakref in FitsContext to work around
the caching.

recurse in schema.walk_schema appears difficult to collect
(perhaps due to the extensive recursion) this is called from
fits_support._load_from_schema which captures several potentially
large memory footprint objects (the hdulist and tree) in the
callback handed to recurse. Adding del appears to improve the
collection and at least in one test cut the memory in half
from ~2 to ~1 GB.

converting DataModel._ctx to property allowed the model to not
immediately jump to generation 2 for garbage collection

* add regression test for #109