link FITS_rec instances to hdu extensions on save
DataModel converts recarrays to FITS_rec on assignment
these were not correctly linked to FITS extensions on
save resulting in duplication of FITS_rec data on save.

This adds a test for the duplication and some book keeping
to allow linking the converted FITS_recs and extensions
on save.

Astropy creates a 'view' of the FITS_rec when assigned
to a new FITS extension which makes a simple 'is' check
complicated (so the strategy used for ndarray and NDArrayType
could not be used). Instead, the links between FITS_rec
and extensions are tracked (via the FitsContext) which allows
creation of the links without having re-find the hdu
created from the FITS_rec. (The existing strategy is left
in place for ndarray to maintain compatibility with existing
ASDF-in-FITS behavior which does not require the same schema
structure as the DataModels).