remove unused V23ToSkyConverter from jwst.transforms extension 1.0.0
V23ToSkyConverter is supported in the jwst.transforms asdf extension
version 0.7.0 (see 0.7.0 manifest) but is not currently supported in
1.0.0 (see 1.0.0 manifest)

This removes V23ToSkyConverter from the 1.0.0 extension. Previously
asdf ignored any Converter that lists tags that are not in the
manifest. Newer versions of asdf will produce errors for Converters
that don't support any supported tags (one's listed in the manifest)
and implement a select_tag method (which this one does by inheriting
from Converter).